requirements.txt
absl-py==2.3.1
alembic==1.16.5
altgraph==0.17.5
annotated-doc==0.0.3
annotated-types==0.7.0
astunparse==1.6.3
Automat==25.4.16
automium==0.2.6
automium_web==0.1.1
cachetools==6.2.1
cattrs==25.3.0
chardet==5.2.0
chex==0.1.90
clearml==2.0.2
coloredlogs==15.0.1
Comet==3.1.0
comet_ml==3.53.2
common==0.1.2
config==0.5.1
configobj==5.0.9
constantly==23.10.4
contourpy==1.3.0
coremltools==8.3.0
cycler==0.12.1
dnspython==2.7.0
dominate==2.9.1
dulwich==0.24.6
email-validator==2.0.0
etils==1.5.2
everett==3.1.0
fastapi==0.121.0
filelock==3.19.1
Flask-Bootstrap==3.3.7.1
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Moment==1.0.6
Flask-MySQLdb==2.0.0
flask-sock==0.7.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==25.9.23
flax==0.8.5
fonttools==4.60.1
FreeSimpleGUI==5.2.0.post1
fsspec==2025.9.0
furl==2.1.4
gast==0.4.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.74.0
h5py==3.14.0
humanfriendly==10.0
humanize==4.13.0
hyperlink==21.0.0
hyperlpr3==0.1.3
idna==3.11
ImageIO==2.37.2
imgviz==1.7.6
importlib_resources==6.5.2
imutils==0.5.4
incremental==24.7.2
jax==0.4.30
jaxlib==0.4.30
joblib==1.5.2
keras==2.15.0
kiwisolver==1.4.7
labelme==5.2.1
lazy_loader==0.4
legacy==0.1.7
libclang==18.1.1
loguru==0.7.3
lxml==6.0.2
Mako==1.3.10
Markdown==3.9
markdown-it-py==3.0.0
matplotlib==3.9.4
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.1.2
mysql==0.0.3
mysqlclient==2.2.7
natsort==8.4.0
networkx==3.2.1
numpy==1.23.5
nvidia-pyindex==1.0.9
oauthlib==3.3.1
onnx==1.19.1
onnx-simplifier==0.4.36
onnxruntime==1.19.2
opencv-python==4.12.0.88
opt_einsum==3.4.0
optax==0.2.4
orbax-checkpoint==0.6.4
orderedmultidict==1.0.1
osam==0.2.5
packaging==23.2
pandas==2.3.3
pathlib2==2.3.7.post1
pefile==2024.8.26
pillow==11.3.0
polars==1.34.0
polars-runtime-32==1.34.0
protobuf==4.25.8
pyaml==25.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.10
pydantic==2.12.4
pydantic_core==2.41.5
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
PyJWT==2.10.1
PyMySQL==1.1.2
pyparsing==3.2.5
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.1
pyreadline3==3.5.4
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
python-box==6.1.0
python-multipart==0.0.20
pytz==2025.2
pywin32-ctypes==0.2.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
scikit-image==0.24.0
scikit-learn==1.1.2
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
sentry-sdk==2.42.1
shiboken6==6.10.1
simple-websocket==1.1.0
simplejson==3.20.2
smmap==5.0.2
sock==0.6.0
SQLAlchemy==2.0.44
starlette==0.49.3
sympy==1.14.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.1
tensorflow-estimator==2.15.0
tensorflow-hub==0.16.1
tensorflow-intel==2.15.1
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_decision_forests==1.8.1
tensorflowjs==4.22.0
tensorstore==0.1.69
termcolor==3.1.0
tf_keras==2.15.1
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2024.8.30
toolz==1.1.0
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
Twisted==25.5.0
typing-inspection==0.4.2
tzdata==2025.2
ultralytics==8.3.217
ultralytics-thop==2.0.17
uvicorn==0.38.0
val==0.6
visitor==0.1.3
waitress==3.0.2
webencodings==0.5.1
websocket-client==0.57.0
win32_setctime==1.2.0
wrapt==1.14.2
wsproto==1.2.0
WTForms==3.2.1
wurlitzer==3.1.1
zope.interface==8.0.1
